Finished documenting RDM state machine.
