+++ Magitek/Utilities/Routines/StateMachine/State.cs
@@ -2,6 +2,9 @@
 
 namespace Magitek.Utilities.Routines
 {
+    //Represents a state. Each state consists simply of an ordered list of state transitions.
+    //See StateTransition.cs for details on how the state machine operates.
+    //See Logic/RedMage/StateMachine.cs for an example of this in action.
     public class State<T>
     {
         public List<StateTransition<T>> Transitions;
